pyproject.toml
[project]
name = "kesson-space-mcp"
version = "0.1.0"
description = "MCP servers for kesson-space Three.js development"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-generativeai>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
